misc(Customer): merge single object and object types
